% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{riem_measures}
\alias{riem_measures}
\title{Function for getting weather data from one station}
\usage{
riem_measures(station = "VOHY", date_start = "2014-01-01",
  date_end = "2016-04-22")
}
\arguments{
\item{station}{station ID, see riem_stations()}

\item{date_start}{date of start of the desired data, e.g. "2000-01-01"}

\item{date_end}{date of end of the desired data, e.g. "2016-04-22"}
}
\value{
a data.frame (dplyr tbl_df) with measures, the number of columns can vary from station to station.
}
\description{
Function for getting weather data from one station
}
\examples{
riem_measures(station = "VOHY", date_start = "2000-01-01", date_end = "2016-04-22")
}

