% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{riem_networks}
\alias{riem_networks}
\title{Function for getting ASOS and AWOS networks}
\usage{
riem_networks()
}
\value{
a data.frame (dplyr tbl_df) with the names and codes of available networks.
}
\description{
Function for getting ASOS and AWOS networks
}
\examples{
riem_networks()
}

