% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{riem_climate}
\alias{riem_climate}
\title{Function for getting climate data from one place or state}
\usage{
riem_climate(the_state = "MA", the_place = "",
  the_date = as.character(Sys.Date()))
}
\arguments{
\item{the_state}{two-letter abbreviation for U.S. state, such as "MA" for Massachusetts or "NY" for New York. See https://mesonet.agron.iastate.edu/COOP/extremes.php for available states and stations, since not all U.S. states are available.}

\item{the_place}{Name of a city or town station name in the U.S. that has an available station ID. Leave as default "" to return entire state's results.}

\item{the_date}{date of desired records in yyyy-mm-dd or mm-dd format, e.g. "2016-04-22" or "04-22".}
}
\description{
Function for getting climate data from one place or state
}
\details{
The data is queried through \url{https://mesonet.agron.iastate.edu/COOP/extremes.php}.
}
\examples{
\dontrun{
riem_climate(the_place = "Boston area", the_state = "MA")
}
}
