% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{riem_networks}
\alias{riem_networks}
\title{Get ASOS and AWOS networks}
\usage{
riem_networks()
}
\value{
a data.frame (tibble tibble) with the names and codes of available networks.
}
\description{
Get ASOS and AWOS networks
}
\examples{
\dontrun{
riem_networks()
}
}
