% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{riem_stations}
\alias{riem_stations}
\title{Function for getting stations of an ASOS network}
\usage{
riem_stations(network = NULL)
}
\arguments{
\item{network}{A single network code, see riem_networks() for finding the code corresponding to a name.}
}
\value{
a data.frame (dplyr tbl_df) with the id, name, longitude (lon) and latitude (lat) of each station in the network.
}
\description{
Function for getting stations of an ASOS network
}
\details{
You can see a map of stations in a network at \url{https://mesonet.agron.iastate.edu/request/download.phtml}.
}
\examples{
\dontrun{
riem_stations(network = "IN__ASOS")
}
}

