% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{riem_stations}
\alias{riem_stations}
\title{Function for getting stations of an ASOS network}
\usage{
riem_stations(network = NULL)
}
\value{
a data.frame (dplyr tbl_df) with the id, name, longitude (lon) and latitude (lat) of each station in the network.
}
\description{
Function for getting stations of an ASOS network
}
\examples{
riem_stations(network = "IN__ASOS")
}

